requirements.txt
python-dotenv~=1.0.0
tiktoken==0.7.0
anthropic==0.45.2
python-telegram-bot==21.9
requests~=2.32.3
tenacity==8.3.0
